requirements.txt
dspy-ai
python-dotenv
git+https://github.com/lordamp/kwansi.git